Add an API to access the packed size of a given type.
This makes it easier to write generic code that uses gopack under the
hood without having to hardcode the size of the structs being used by
the generic code.

This change also removes a layer of indirection in makePackerWrapper,
by having the code check the input buffer is big enough directly in
Pack() rather than via an extra anonymous func, which improves most
encoding benchmarks by 4 to 6% (the gains are really modest in absolute
terms, on the order of a few cycles, as all it's doing really is just
removing an extra indirect function call).